backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
gunicorn>=21.2.0
jinja2>=3.1.0
jieba>=0.42.1
openai>=1.55.3
httpx>=0.27.0
playwright>=1.40.0
pymysql>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
ijson>=3.2.0
